remove dependencies on makefile
If the source properties depend are defined in the makefile itself,
several targets (those which use MOS_SRC etc.) should depend on
the makefile itself. However, during development and debugging this
would cause too much reprocessing. so, I removed these dependencies for now
and put a note in the makefile.